python/requirements.txt
flask==2.2.3
pandas==1.5.0
sodapy==2.2.0
boto3==1.26.133
python-dotenv==1.0.0